% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.R
\docType{package}
\name{agriPairMetaFlow-package}
\alias{agriPairMetaFlow}
\alias{agriPairMetaFlow-package}
\title{agriPairMetaFlow: treatment-control meta-analysis for agronomy}
\description{
\code{agriPairMetaFlow} provides an estimand-first, design-first interface for
treatment-versus-control evidence synthesis. Version 0.4.0 adds joint
multivariate outcome synthesis and Bayesian workflows while preserving the
dependence-aware, dose-response, meta-regression, prediction, and plotting
layers introduced in earlier versions. Backend objects remain available for
auditability and local numerical validation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wep69/agriPairMetaFlow}
  \item Report bugs at \url{https://github.com/wep69/agriPairMetaFlow/issues}
}

}
\author{
\strong{Maintainer}: Walter Esfrain Pereira \email{walterufpb@yahoo.com.br} (\href{https://orcid.org/0000-0003-1085-0191}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Walter Esfrain Pereira \email{walterufpb@yahoo.com.br} (\href{https://orcid.org/0000-0003-1085-0191}{ORCID}) [copyright holder]
  \item Magali Haidee Pereira Martinez (\href{https://orcid.org/0009-0009-5419-959X}{ORCID})
}

}
\keyword{internal}
